feat: update metrics to easily see sync status
Updates the metrics dashboard so users can more easily see the status of
the node and can actually figure out when the node is ready to be used.
Once the PR is merged to allow Geth to sync alongside the DTL, I'll
likely also move the DTL sync panel into a separate row since it's a
little confusing and doesn't really tell you much.