progress: Avoid accidental early return
expr returns a non-zero exit code if it's calculated value is zero. This
is surprising and usually not what you want. In this case, it will lead
to the script terminating early for some values (e.g. HOURS=0). Small
example to show the problem:

```
> bash -c 'HOURS=`expr 0` && echo true || echo false'
false
```

Fortunately, bash has the arithmetic calculation syntax $((...)) that
behaves as intended.