Remove explicit selection of pebble db (#17)
This allows for use of a pre-exising datadir with leveldb.

If the db.engine flag is explicitly set then only the chose db engine
can be used. If the db.engine flag is unset then it will first use the db
engine found in the datadir, and if there is no datadir it will default to pebble.